% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{categorizeData}
\alias{categorizeData}
\title{Categorize Data}
\usage{
categorizeData(df, cohorts)
}
\arguments{
\item{df}{A dataframe with samples in columns}

\item{cohorts}{A character vector of cohort identifiers}
}
\value{
A list of dataframes: first element contains character columns, subsequent elements contain numeric columns for each cohort
}
\description{
Splits a dataframe into multiple dataframes based on cohort identifiers in column names.
}
\details{
Categorize Data by Cohorts
}
\examples{
# df <- data.frame(Gene=c("A", "B"), Control_1=c(1,2), Control_2=c(3,4), Treated_1=c(5,6))
# categorized <- categorizeData(df, c("Control", "Treated"))
}
