% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compareCohorts}
\alias{compareCohorts}
\title{Compare Cohorts}
\usage{
compareCohorts(data_list, test = "mwtest")
}
\arguments{
\item{data_list}{A list of dataframes containing categorized data}

\item{test}{The statistical test to use: "mwtest" (Mann-Whitney) or "ttest" (t-test)}
}
\value{
A list of dataframes containing comparison results
}
\description{
Performs pairwise statistical comparisons between all cohorts in a list.
}
\details{
Compare Cohorts with Statistical Tests
}
\examples{
# data_list <- list(df_char, df_cohort1, df_cohort2)
# results <- compareCohorts(data_list, test = "mwtest")
}
