% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extractGeneName}
\alias{extractGeneName}
\title{Extract Gene Name}
\usage{
extractGeneName(description)
}
\arguments{
\item{df}{A dataframe containing a 'Description' column}
}
\value{
A dataframe with added 'Gene' and 'Protein' columns
}
\description{
Given a dataframe from Proteome Discoverer with a 'Description' column, extracts the gene name
from the GN= field and the protein name.
}
\details{
Extract Gene Name and Protein Name from Protein Description
}
\examples{
# df <- data.frame(Description = "Protein OS=Homo sapiens GN=ACTB PE=1 SV=1")
# df <- extractGeneName(df)
}
