% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filterHighAbundance}
\alias{filterHighAbundance}
\title{Filter High Abundance}
\usage{
filterHighAbundance(df)
}
\arguments{
\item{df}{A dataframe containing an 'Accession' column}
}
\value{
A filtered dataframe with high abundance proteins removed
}
\description{
Removes common high abundance proteins (albumin, immunoglobulins, etc.) from the dataset.
}
\details{
Filter High Abundance Proteins
}
\examples{
# df <- data.frame(Accession=c("P02768", "Q12345"), value=c(100, 50))
# filtered <- filterHighAbundance(df)
}
