% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{filterKeratin}
\alias{filterKeratin}
\title{Filter Keratin}
\usage{
filterKeratin(df)
}
\arguments{
\item{df}{A dataframe containing a 'Description' column}
}
\value{
A filtered dataframe with keratin proteins removed
}
\description{
Removes keratin proteins (common contaminants) from the dataset.
}
\details{
Filter Keratin Proteins
}
\examples{
# df <- data.frame(Description=c("Keratin type I", "Actin"), value=c(100, 50))
# filtered <- filterKeratin(df)
}
