% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performGo.R
\name{geneOntology}
\alias{geneOntology}
\title{Gene Ontology Analysis}
\usage{
geneOntology(df, ontology = "BP", database = org.Rn.eg.db, keytype = "SYMBOL")
}
\arguments{
\item{df}{A dataframe containing a 'Gene' column with gene symbols}

\item{ontology}{The ontology to use: "BP" (Biological Process), "CC" (Cellular Component), or "MF" (Molecular Function)}

\item{database}{The organism database to use (default: org.Rn.eg.db for rat)}

\item{keytype}{The type of gene identifier (default: "SYMBOL")}
}
\value{
A dataframe containing GO enrichment results
}
\description{
Performs gene ontology enrichment analysis on a set of genes using clusterProfiler.
}
\details{
Perform Gene Ontology Analysis
}
\examples{
# df <- data.frame(Gene = c("ACTB", "GAPDH", "TP53"))
# go_results <- geneOntology(df, ontology = "BP")
}
