% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{importData}
\alias{importData}
\title{Import Data}
\usage{
importData(filepath)
}
\arguments{
\item{filepath}{Path to the file to import}
}
\value{
A dataframe containing the imported data
}
\description{
Imports data from various file formats (.csv, .xlsx, .xls).
}
\details{
Import Data from File
}
\examples{
# data <- importData("data.csv")
}
