% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/utilities_old.R
\name{imputeValues}
\alias{imputeValues}
\title{imputeValues}
\usage{
imputeValues(row)

imputeValues(row)
}
\arguments{
\item{row}{a row from a dataframe containing missing values}
}
\value{
a row with missing values imputed using truncated normal distribution

a row with missing values imputed using truncated normal distribution
}
\description{
imputeValues

imputeValues
}
\examples{
imputeValues(df[1, ])
imputeValues(df[1, ])
}
