% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeBarplot3}
\alias{makeBarplot3}
\title{makeBarplot3}
\usage{
makeBarplot3(df, pvalue_df, cohort_labels, gene)
}
\arguments{
\item{df}{long dataframe containing the values}

\item{pvalue_df}{the main wide dataframe containing the pvalues from the ANOVA result}

\item{cohort_labels}{vector containing the cohort labels}

\item{gene}{the gene to be plotted}
}
\value{
ggplot2 object containing the barplot
}
\description{
makeBarplot3
}
\examples{
\dontrun{
  df_long <- data.frame(
    Treatment = rep(c("A", "B", "C"), each = 3),
    Abundance = rnorm(9)
  )
  tukey_df <- data.frame(
    Gene = "Gene1",
    tukey_pvalue12 = 0.03,
    tukey_pvalue13 = 0.2,
    tukey_pvalue23 = 0.04
  )
  makeBarplot3(df_long, tukey_df, c("A", "B", "C"), "Gene1")
}
}
