% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBarplot_simp.R
\name{makeBarplot_simp}
\alias{makeBarplot_simp}
\title{Make Simple Barplot}
\usage{
makeBarplot_simp(df, cohort_labels, gene, num = 0)
}
\arguments{
\item{df}{A long-format dataframe with 'Treatment' and 'Abundance' columns}

\item{cohort_labels}{A vector of cohort labels}

\item{gene}{The gene name to plot}

\item{num}{Additional space to add to y-axis maximum (default 0)}
}
\value{
A ggplot2 object containing the barplot
}
\description{
Creates a simple barplot without statistical annotations.
}
\details{
Create Simple Barplot
}
\examples{
# df_long <- data.frame(Treatment=rep(c("A","B"),each=5), Abundance=rnorm(10))
# plot <- makeBarplot_simp(df_long, c("A","B"), "ACTB")
}
