% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/utilities_old.R
\name{performANOVA4}
\alias{performANOVA4}
\title{performANOVA4}
\usage{
performANOVA4(df1, df2, df3, df4)

performANOVA4(df1, df2, df3, df4)
}
\arguments{
\item{df1}{dataframe containing cohort 1}

\item{df2}{dataframe containing cohort 2}

\item{df3}{dataframe containing cohort 3}

\item{df4}{dataframe containing cohort 4}
}
\value{
A dataframe with the results of the ANOVA

A dataframe with the results of the ANOVA
}
\description{
Takes in 4 dataframes, each belonging to a cohort, and performs an ANOVA on each row.
The function returns a dataframe with the p-value from the ANOVA, Tukey's HSD p-values for pairwise
comparisons, and fold changes for each pairwise comparison. Only meant to be used with compareCohorts.

Takes in 4 dataframes, each belonging to a cohort, and performs an ANOVA on each row.
The function returns a dataframe with the p-value from the ANOVA, Tukey's HSD p-values for pairwise
comparisons, and fold changes for each pairwise comparison. Only meant to be used with compareCohorts.
}
\examples{
test <- performANOVA4(df1, df2, df3, df4)
test <- performANOVA4(df1, df2, df3, df4)
}
