% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/utilities_old.R
\name{performANOVA6}
\alias{performANOVA6}
\title{performANOVA6}
\usage{
performANOVA6(df1, df2, df3, df4, df5, df6)

performANOVA6(df1, df2, df3, df4, df5, df6)
}
\arguments{
\item{df1}{Data frame containing cohort 1}

\item{df2}{Data frame containing cohort 2}

\item{df3}{Data frame containing cohort 3}

\item{df4}{Data frame containing cohort 4}

\item{df5}{Data frame containing cohort 5}

\item{df6}{Data frame containing cohort 6}
}
\value{
A data frame with the results of the ANOVA, Tukey’s HSD p-values, and fold changes

A data frame with the results of the ANOVA, Tukey’s HSD p-values, and fold changes
}
\description{
Takes in 6 data frames, each belonging to a cohort, and performs an ANOVA on each row.
The function returns a data frame with the p-value from the ANOVA, Tukey's HSD p-values
for pairwise comparisons, and fold changes for each pairwise comparison.

Takes in 6 data frames, each belonging to a cohort, and performs an ANOVA on each row.
The function returns a data frame with the p-value from the ANOVA, Tukey's HSD p-values
for pairwise comparisons, and fold changes for each pairwise comparison.
}
\examples{
test <- performANOVA6(df1, df2, df3, df4, df5, df6)
test <- performANOVA6(df1, df2, df3, df4, df5, df6)
}
