% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{performGO}
\alias{performGO}
\title{Perform GO Analysis on List}
\usage{
performGO(data)
}
\arguments{
\item{data}{A list of dataframes, each containing a 'Gene' column}
}
\value{
A list of lists containing GO results for each dataframe and ontology
}
\description{
Takes a list of dataframes and performs gene ontology analysis on each,
across all three ontologies (BP, MF, CC).
}
\details{
Perform Gene Ontology on Multiple Dataframes
}
\examples{
# df1 <- data.frame(Gene = c("ACTB", "GAPDH"))
# df2 <- data.frame(Gene = c("TP53", "MYC"))
# results <- performGO(list(df1, df2))
}
