% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{performMWTest}
\alias{performMWTest}
\title{Row-wise Mann-Whitney Test}
\usage{
performMWTest(df1, df2)
}
\arguments{
\item{df1}{Dataframe containing the first cohort}

\item{df2}{Dataframe containing the second cohort}
}
\value{
A dataframe with the original data plus pvalue, padj, foldchange, and nlog10p columns
}
\description{
Performs a Mann-Whitney U test on each row between two dataframes.
}
\details{
Perform Mann-Whitney Test on Two Dataframes
}
\examples{
# df1 <- data.frame(s1=rnorm(10), s2=rnorm(10), s3=rnorm(10))
# df2 <- data.frame(s4=rnorm(10), s5=rnorm(10), s6=rnorm(10))
# results <- performMWTest(df1, df2)
}
