% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicsUtils-package.R
\docType{package}
\name{proteomicsUtils-package}
\alias{proteomicsUtils}
\alias{proteomicsUtils-package}
\title{proteomicsUtils: Proteomics Data Analysis Utilities}
\description{
A collection of utility functions for proteomics data analysis, including data import, normalization, filtering, statistical testing, visualization (volcano plots, heatmaps, PCA), gene ontology analysis, and machine learning utilities.
}
\author{
\strong{Maintainer}: Vishal \email{vishal.sandilya@ttu.edu}

}
\keyword{internal}
