% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extractGeneName}
\alias{extractGeneName}
\title{Extract Gene Name and Protein Name}
\usage{
extractGeneName(df)
}
\arguments{
\item{df}{Data frame containing a 'Description' column}
}
\value{
Data frame with added 'Gene' and 'Protein' columns
}
\description{
Given a dataframe of PD Result with a 'Description' column, this function extracts the
Gene Name and Protein Name
}
\details{
Extract Gene Name and Protein Name from Description column
}
\examples{
df <- data.frame(Description = c("sp|P12345|PROT1_HUMAN Protein Name 1 GN=GENE1",
                                 "sp|P67890|PROT2_HUMAN Protein Name 2 GN=GENE2"))
df_extracted <- extract_gene_protein(df)
}
