% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extractGeneName}
\alias{extractGeneName}
\title{Extract Gene Name}
\usage{
extractGeneName(description)
}
\arguments{
\item{description}{A character string containing the protein description}
}
\value{
A character string containing the gene name
}
\description{
Given a protein description from Proteome Discoverer, extracts the gene name
from the GN= field.
}
\details{
Extract Gene Name from Protein Description
}
\examples{
# desc <- "Protein OS=Homo sapiens GN=ACTB PE=1 SV=1"
# gene <- extractGeneName(desc)
}
