% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{performTTest}
\alias{performTTest}
\title{Row-wise T-Test}
\usage{
performTTest(df1, df2)
}
\arguments{
\item{df1}{Dataframe containing the first cohort}

\item{df2}{Dataframe containing the second cohort}
}
\value{
A dataframe with the original data plus pvalue, padj, foldchange, nlog10p, and log2fc columns
}
\description{
Performs a t-test on each row between two dataframes and calculates
fold changes and adjusted p-values.
}
\details{
Perform T-Test on Two Dataframes
}
\examples{
# df1 <- data.frame(s1=rnorm(10), s2=rnorm(10), s3=rnorm(10))
# df2 <- data.frame(s4=rnorm(10), s5=rnorm(10), s6=rnorm(10))
# results <- performTTest(df1, df2)
}
